# api/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Flight Control API
  version: v1alpha1
  description: |
    [Flight Control](https://flightctl.io) is a service for declarative management of fleets of edge devices and their workloads.
  contact:
    name: The Flight Control Team
    url: https://flightctl.io
    email: team@flightctl.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/v1/organizations/{orgID}/resourcesyncs:
    get:
      tags:
        - resourcesync
      description: List ResourceSync resources.
      operationId: listResourceSyncs
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: continue
          in: query
          description: An optional parameter to query more results from the server. The value of the paramter must match the value of the 'continue' field in the previous list response.
          required: false
          schema:
            type: string
        - name: labelSelector
          in: query
          description: A selector to restrict the list of returned objects by their labels. Defaults to everything.
          schema:
            type: string
        - name: fieldSelector
          in: query
          description: A selector to restrict the list of returned objects by their fields, supporting operators like '=', '==', and '!=' (e.g., "key1=value1,key2!=value2").
          schema:
            type: string
        - name: limit
          in: query
          description: The maximum number of results returned in the list response. The server will set the 'continue' field in the list response if more results exist. The continue value may then be specified as parameter in a subsequent query.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSyncList'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    post:
      tags:
        - resourcesync
      description: Create a ResourceSync resource.
      operationId: createResourceSync
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceSync'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSync'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: StatusConflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  /api/v1/organizations/{orgID}/resourcesyncs/{name}:
    get:
      tags:
        - resourcesync
      description: Get a ResourceSync resource.
      operationId: getResourceSync
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the ResourceSync resource to get.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSync'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    put:
      tags:
        - resourcesync
      description: Update a ResourceSync resource.
      operationId: replaceResourceSync
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the ResourceSync resource to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceSync'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSync'
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSync'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    patch:
      tags:
        - resourcesync
      description: Patch a ResourceSync resource.
      operationId: patchResourceSync
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the ResourceSync resource to patch.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/PatchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceSync'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      tags:
        - resourcesync
      description: Delete a ResourceSync resource.
      operationId: deleteResourceSync
      parameters:
        - name: orgID
          in: path
          description: The ID of the organization.
          required: true
          schema:
            type: string
            format: uuid
        - name: name
          in: path
          description: The name of the ResourceSync resource to delete.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        "503":
          description: ServiceUnavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
components:
  schemas:
...
    PatchRequest:
      type: array
      items:
        type: object
        additionalProperties: false
        required:
          - op
          - path
        properties:
          path:
            description: A JSON Pointer path.
            type: string
          value:
            description: The value to add or replace.
          op:
            description: The operation to perform.
            type: string
            enum:
              - add
              - replace
              - remove
...
    ResourceSync:
      type: object
      description: ResourceSync represents a reference to one or more files in a repository to sync to resource definitions.
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ObjectMeta'
        spec:
          $ref: '#/components/schemas/ResourceSyncSpec'
        status:
          $ref: '#/components/schemas/ResourceSyncStatus'
      required:
        - apiVersion
        - kind
        - metadata
        - spec
    ResourceSyncSpec:
      type: object
      description: ResourceSyncSpec describes the file(s) to sync from a repository.
      properties:
        repository:
          type: string
          description: The name of the repository resource to use as the sync source.
        targetRevision:
          type: string
          description: The desired revision in the repository.
        path:
          type: string
          description: The path of a file or directory in the repository. If a directory, the directory should contain only resource definitions with no subdirectories. Each file should contain the definition of one or more resources.
      required:
      - repository
      - targetRevision
      - path
    ResourceSyncStatus:
      type: object
      properties:
        observedCommit:
          type: string
          description: The last commit hash that was synced.
        observedGeneration:
          type: integer
          format: int64
          description: The last generation that was synced.
        conditions:
          type: array
          description: Current state of a resourcesync.
          items:
            $ref: '#/components/schemas/Condition'
      required:
        - conditions
      description: ResourceSyncStatus represents information about the status of a ResourceSync.
    ResourceSyncList:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        metadata:
          $ref: '#/components/schemas/ListMeta'
        items:
          type: array
          description: 'List of resourcesync.'
          items:
            $ref: '#/components/schemas/ResourceSync'
      required:
        - apiVersion
        - kind
        - metadata
        - items
...
    ListMeta:
      type: object
      properties:
        continue:
          type: string
          description: May be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.
        remainingItemCount:
          type: integer
          description: The number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.
          format: int64
      description: ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.
    ObjectMeta:
      type: object
      properties:
        creationTimestamp:
          type: string
          description: The time the object was created.
          format: date-time
        deletionTimestamp:
          type: string
          description: The time the object will be deleted.
          format: date-time
        name:
          type: string
          description: The name of the object.
        labels:
          type: object
          description: Map of string keys and values that can be used to organize and categorize (scope and select) objects.
          additionalProperties:
            type: string
        generation:
          type: integer
          description: A sequence number representing a specific generation of the desired state. Populated by the system. Read-only.
          format: int64
        owner:
          type: string
          description: A resource that owns this resource, in "kind/name" format.
        annotations:
          type: object
          additionalProperties:
            type: string
          description: Properties set by the service.
        resourceVersion:
          type: string
          description: An opaque string that identifies the server's internal version of an object.
      description: ObjectMeta is metadata that all persisted resources must have, which includes all objects users must create.
...
    Status:
      type: object
      properties:
        apiVersion:
          type: string
          description: 'APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources.'
        kind:
          type: string
          description: 'Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds.'
        code:
          type: integer
          format: int32
          description: Suggested HTTP return code for this status, 0 if not set.
        message:
          type: string
          description: A human-readable description of the status of this operation.
        reason:
          type: string
          description: A machine-readable description of why this operation is in the "Failure" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.
        status:
          type: string
          description: 'Status of the operation. One of: "Success" or "Failure". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status.'
      description: Status is a return value for calls that don't return other objects.
      required:
        - apiVersion
        - kind
        - status
        - code
        - reason
        - message
    Condition:
      required:
        - type
        - status
        - lastTransitionTime
        - reason
        - message
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ConditionType'
        status:
          $ref: '#/components/schemas/ConditionStatus'
        observedGeneration:
          type: integer
          format: int64
          description: The .metadata.generation that the condition was set based upon.
        lastTransitionTime:
          type: string
          format: date-time
          description: The last time the condition transitioned from one status to another.
        message:
          type: string
          description: Human readable message indicating details about last transition.
        reason:
          type: string
          description: A (brief) reason for the condition's last transition.
      description: Condition contains details for one aspect of the current state of this API Resource.
    ConditionType:
      type: string
      description: Type of condition in CamelCase.
      enum:
      - 'Approved'              # EnrollmentRequest
      - 'Approved'              # CertificateSigningRequest
      - 'Denied'                # CertificateSigningRequest
      - 'Failed'                # CertificateSigningRequest
      - 'Accessible'            # Repository
      - 'Accessible'            # ResourceSync
      - 'ResourceParsed'        # ResourceSync
      - 'Synced'                # ResourceSync
      - 'OverlappingSelectors'  # Fleet
      - 'Valid'                 # Fleet
      - 'RolloutInProgress'     # Fleet
      - 'Updating'              # Device
      - 'SpecValid'             # Device (service condition)
      - 'MultipleOwners'        # Device (service condition)
      - 'DeviceDecommissioning' # Device
      x-enum-varnames:
      - EnrollmentRequestApproved
      - CertificateSigningRequestApproved
      - CertificateSigningRequestDenied
      - CertificateSigningRequestFailed
      - RepositoryAccessible
      - ResourceSyncAccessible
      - ResourceSyncResourceParsed
      - ResourceSyncSynced
      - FleetOverlappingSelectors
      - FleetValid
      - FleetRolloutInProgress
      - DeviceUpdating
      - DeviceSpecValid
      - DeviceMultipleOwners
      - DeviceDecommissioning
    ConditionStatus:
      type: string
      description: Status of the condition, one of True, False, Unknown.
      enum:
      - 'True'
      - 'False'
      - 'Unknown'
      x-enum-varnames:
      - 'ConditionStatusTrue'
      - 'ConditionStatusFalse'
      - 'ConditionStatusUnknown'
...